added feedback and removed some files
